Minor changes with security and added more bussiness logic in Service

+++ src/main/java/payoya/com/service/PostService.java
@@ -33,7 +33,12 @@ public class PostService {
         return postRepository.findById(id).orElse(null);
     }
 
-    public void deleteById(Long id) {
-        postRepository.deleteById(id);
+    public boolean deleteById(Long id) {
+        if(postRepository.existsById(id)){
+            postRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
